refactor: simple adjustments of tetris
