+++ app/src/main/java/com/iccas/zen/presentation/components/PlayControlButton.kt
@@ -1,4 +1,4 @@
-package com.iccas.zen.presentation.tetris.tetrisComponents
+package com.iccas.zen.presentation.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -20,21 +19,21 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-
 
 @Composable
 fun PlayControlButton(
+    modifier: Modifier = Modifier,
     onClick: () -> Unit,
     background: Color,
     imageVector: ImageVector,
     iconColor: Color,
+    iconSize: Dp,
     contentDescription: String
 ) {
     Box(
-        modifier = Modifier
-            .width(110.dp)
-            .height(50.dp)
+        modifier = modifier
             .background(background, RoundedCornerShape(50))
             .clickable { onClick() },
         contentAlignment = Alignment.Center
@@ -42,8 +41,7 @@ fun PlayControlButton(
         Icon(
             imageVector = imageVector,
             contentDescription = contentDescription,
-            modifier = Modifier
-                .size(30.dp),
+            modifier = Modifier.size(iconSize),
             tint = iconColor
         )
     }
@@ -51,12 +49,14 @@ fun PlayControlButton(
 
 @Composable
 fun ReplayAndExitControlButtons(
+    modifier: Modifier = Modifier,
     onReplay: () -> Unit,
     onExit: () -> Unit,
     replayButtonBackground: Color,
     exitButtonBackground: Color,
     replayIconColor: Color,
-    exitIconColor: Color
+    exitIconColor: Color,
+    iconSize: Dp
 ) {
     Row(
         modifier = Modifier.fillMaxWidth(),
@@ -64,20 +64,24 @@ fun ReplayAndExitControlButtons(
         verticalAlignment = Alignment.CenterVertically
     ) {
         PlayControlButton(
+            modifier = modifier,
             onClick = { onReplay() },
             background = replayButtonBackground,
             imageVector = Icons.Filled.Refresh,
             contentDescription = "Replay",
-            iconColor = replayIconColor
+            iconColor = replayIconColor,
+            iconSize = iconSize
         )
         Spacer(modifier = Modifier.width(15.dp))
 
         PlayControlButton(
+            modifier = modifier,
             onClick = { onExit() },
             background = exitButtonBackground,
             imageVector = Icons.Filled.ExitToApp,
             contentDescription = "Exit",
-            iconColor = exitIconColor
+            iconColor = exitIconColor,
+            iconSize = iconSize
         )
     }
 }
